+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# Dodo x.x.x (not yet released)
+# Change Log
+
+## Dodo x.x.x (not yet released)
 * Update to wikimedia/zest-css 2.0.0.
 * Implement the following non-standard methods for PHP compatibility:
   * Document::loadXML()
@@ -6,7 +8,7 @@
   * DocumentFragment::appendXML()
   * Element::setIdAttribute() (stubbed out)
 
-# Dodo 0.2.0 (2021-07-26)
+## Dodo 0.2.0 (2021-07-26)
 * Update to IDLeDOM 0.7.2.
 * Fix doctype creation and "significant whitespace" handling in DOMParser.
 * Ensure that Document::documentElement is always populated.
@@ -20,5 +22,5 @@
 * Implement Node::getNodePath() for PHP compatibility.
 * Don't export .phan directory in composer package.
 
-# Dodo 0.1.0 (2021-07-04)
+## Dodo 0.1.0 (2021-07-04)
 Initial release.